refactor: 클래스 별도 파일로 분리 및 네이밍 수정
