+++ config-server/src/main/kotlin/com/gitub/oopgurus/refactoringproblems/configserver/WhatIWantToProperties.kt
@@ -1,0 +1,7 @@
+package com.gitub.oopgurus.refactoringproblems.configserver
+
+interface WhatIWantToProperties {
+    fun descriptions(descriptions: List<String>)
+
+    fun properties(properties: Map<String, String>)
+}
